Move HSolar, Wind in file
