+++ src/libs/antares/study/scenario-builder/TSnumberData.h
@@ -122,38 +122,6 @@ inline CString<512, false> loadTSNumberData::get_prefix() const
     return "l,";
 }
 
-// =====================
-// Wind ...
-// =====================
-class windTSNumberData : public TSNumberData
-{
-public:
-    bool apply(Study& study) override;
-    CString<512, false> get_prefix() const override;
-    uint get_tsGenCount(const Study& study) const override;
-};
-
-inline CString<512, false> windTSNumberData::get_prefix() const
-{
-    return "w,";
-}
-
-// =====================
-// Solar ...
-// =====================
-class solarTSNumberData : public TSNumberData
-{
-public:
-    bool apply(Study& study) override;
-    CString<512, false> get_prefix() const override;
-    uint get_tsGenCount(const Study& study) const override;
-};
-
-inline CString<512, false> solarTSNumberData::get_prefix() const
-{
-    return "s,";
-}
-
 } // namespace Antares::Data::ScenarioBuilder
 
 #endif // __LIBS_STUDY_SCENARIO_BUILDER_DATA_TS_NUMBER_H__
